+++ packages/dart/noports_core/lib/src/sshnp/util/srvd_channel/srvd_channel.dart
@@ -5,11 +5,11 @@ import 'package:at_utils/at_utils.dart';
 import 'package:meta/meta.dart';
 import 'package:noports_core/src/common/mixins/async_initialization.dart';
 import 'package:noports_core/src/common/mixins/at_client_bindings.dart';
-import 'package:noports_core/src/common/validation_utils.dart';
 import 'package:noports_core/src/sshnp/util/srvd_channel/notification_request_message.dart';
 import 'package:noports_core/sshnp.dart';
 import 'package:noports_core/srv.dart';
 import 'package:noports_core/srvd.dart';
+import 'package:noports_core/utils.dart';
 
 @visibleForTesting
 enum SrvdAck {
@@ -98,6 +98,7 @@ abstract class SrvdChannel<T> with AsyncInitialization, AtClientBindings {
     String? sessionIVString,
     bool multi = false,
     bool detached = false,
+    Duration timeout = DefaultArgs.srvTimeout,
   }) async {
     await callInitialization();
 
@@ -122,6 +123,7 @@ abstract class SrvdChannel<T> with AsyncInitialization, AtClientBindings {
       sessionIVString: sessionIVString,
       multi: multi,
       detached: detached,
+      timeout: timeout,
     );
     return srv.run();
   }
@@ -206,7 +208,7 @@ abstract class SrvdChannel<T> with AsyncInitialization, AtClientBindings {
       counter++;
       if (counter > 150) {
         logger.warning('Timed out waiting for srvd response');
-        throw SshnpError(
+        throw TimeoutException(
             'Connection timeout to srvd ${params.srvdAtSign} service');
       }
     }